feat(phase-4): Cypher parser, binder, and factorized execution engine
Implements Phase 4 (4a + 4b + 4c) — makes SparrowDB queryable end-to-end.

- Hand-written lexer (lexer.rs): tokenizes Cypher including keywords, operators, string/int/float/param literals
- Recursive-descent parser (parser.rs): covers CREATE, MATCH, WHERE, RETURN, ORDER BY, LIMIT, DISTINCT, CHECKPOINT, OPTIMIZE; returns Err (never panics) for unsupported syntax (OPTIONAL MATCH, UNION, UNWIND, DETACH DELETE, variable-length paths)
- Full AST (ast.rs): NodePattern, RelPattern, PathPattern, MatchStatement, CreateStatement, MatchCreateStatement, Expr, Literal, ReturnClause, etc.
- Binder (binder.rs): resolves labels and rel-type names against Catalog; returns Error::InvalidArgument for unknown labels or rel types
- Added list_rel_tables() and get_rel_table() to sparrowdb-catalog::Catalog

- TypedVector, VectorGroup, FactorizedChunk, QueryResult (types.rs)
- Operator trait with next_chunk() iterator pattern
- LabelScan: reads all slots for a label_id from NodeStore
- Filter: evaluates Eq / Contains / Gt predicates over FactorizedChunks
- Project: selects named columns from chunks
- Expand: CSR neighbor lookup, multiplicity-preserving
- Added NodeStore::hwm_for_label() to sparrowdb-storage

- AspJoin::two_hop(): Roaring-bitmap semijoin filter + hash state + re-probe
- AspJoin::two_hop_factorized(): factorized form, propagates multiplicity without materializing Cartesian products
- Engine::execute_two_hop(): wires ASP-Join into end-to-end query execution

- Acceptance #3 (single_reader_match_person_names): PASS — 1-hop MATCH friend lookup
- Acceptance #4 (two_hop_match_via_binary_asp_join): PASS — 2-hop fof via binary ASP-Join

- 19 parser tests, 6 binder tests, 6 join unit tests, 3 scan/filter/project tests, 2 UC-1 integration tests

+++ crates/sparrowdb/tests/uc1_social_graph.rs
@@ -4,8 +4,6 @@
 //! Acceptance check #4: two_hop_match_via_binary_asp_join
 
 use sparrowdb_catalog::catalog::Catalog;
-use sparrowdb_cypher::parser::parse;
-use sparrowdb_cypher::binder::bind;
 use sparrowdb_execution::engine::Engine;
 use sparrowdb_storage::csr::CsrForward;
 use sparrowdb_storage::node_store::{NodeStore, Value as StoreValue};
@@ -22,26 +20,41 @@ fn setup_social_graph(dir: &std::path::Path) -> Engine {
     // col 0 = name (stored as i64 id: 1=Alice, 2=Bob, 3=Carol, 4=Dave, 5=Eve)
     // col 1 = age
     let alice = store
-        .create_node(person_id, &[(0, StoreValue::Int64(1)), (1, StoreValue::Int64(30))])
+        .create_node(
+            person_id,
+            &[(0, StoreValue::Int64(1)), (1, StoreValue::Int64(30))],
+        )
         .expect("Alice");
     let bob = store
-        .create_node(person_id, &[(0, StoreValue::Int64(2)), (1, StoreValue::Int64(25))])
+        .create_node(
+            person_id,
+            &[(0, StoreValue::Int64(2)), (1, StoreValue::Int64(25))],
+        )
         .expect("Bob");
     let carol = store
-        .create_node(person_id, &[(0, StoreValue::Int64(3)), (1, StoreValue::Int64(35))])
+        .create_node(
+            person_id,
+            &[(0, StoreValue::Int64(3)), (1, StoreValue::Int64(35))],
+        )
         .expect("Carol");
     let dave = store
-        .create_node(person_id, &[(0, StoreValue::Int64(4)), (1, StoreValue::Int64(28))])
+        .create_node(
+            person_id,
+            &[(0, StoreValue::Int64(4)), (1, StoreValue::Int64(28))],
+        )
         .expect("Dave");
     let _eve = store
-        .create_node(person_id, &[(0, StoreValue::Int64(5)), (1, StoreValue::Int64(22))])
+        .create_node(
+            person_id,
+            &[(0, StoreValue::Int64(5)), (1, StoreValue::Int64(22))],
+        )
         .expect("Eve");
 
     // Edges: Alice->Bob, Alice->Carol, Bob->Dave, Carol->Dave, Carol->Eve
-    let alice_slot = (alice.0 & 0xFFFF_FFFF) as u64;
-    let bob_slot = (bob.0 & 0xFFFF_FFFF) as u64;
-    let carol_slot = (carol.0 & 0xFFFF_FFFF) as u64;
-    let dave_slot = (dave.0 & 0xFFFF_FFFF) as u64;
+    let alice_slot = alice.0 & 0xFFFF_FFFF;
+    let bob_slot = bob.0 & 0xFFFF_FFFF;
+    let carol_slot = carol.0 & 0xFFFF_FFFF;
+    let dave_slot = dave.0 & 0xFFFF_FFFF;
 
     let edges = vec![
         (alice_slot, bob_slot),
